+++ ydb/core/blobstorage/ddisk/ddisk_actor_read_write.cpp
@@ -92,7 +92,7 @@ namespace NKikimr::NDDisk {
                 NWilson::EFlags::NONE, TActivationContext::ActorSystem())
             .Attribute("tablet_id", static_cast<long>(creds.TabletId))
             .Attribute("vchunk_index", static_cast<long>(selector.VChunkIndex))
-            .Attribute("offset_in_bytes", selector.OffsetInBytes)
+            .Attribute("offset_in_bytes", static_cast<long>(selector.OffsetInBytes))
             .Attribute("size", selector.Size));
 
         TRope data;
@@ -160,7 +160,7 @@ namespace NKikimr::NDDisk {
                 NWilson::EFlags::NONE, TActivationContext::ActorSystem())
             .Attribute("tablet_id", static_cast<long>(creds.TabletId))
             .Attribute("vchunk_index", static_cast<long>(selector.VChunkIndex))
-            .Attribute("offset_in_bytes", selector.OffsetInBytes)
+            .Attribute("offset_in_bytes", static_cast<long>(selector.OffsetInBytes))
             .Attribute("size", selector.Size));
 
         if (!chunkRef.ChunkIdx) {
